Migrate from `rustybuzz` and `ttf-parser` to `harfrust` and `fontations`

+++ crates/typst-pdf/src/text.rs
@@ -110,7 +110,7 @@ struct PdfGlyph(Glyph);
 impl krilla::text::Glyph for PdfGlyph {
     #[inline(always)]
     fn glyph_id(&self) -> GlyphId {
-        GlyphId::new(self.0.id as u32)
+        GlyphId::new(self.0.id)
     }
 
     #[inline(always)]
